Refactored ButtonTest per merge from master which refactored ButtonFocusTest. Fixed bug in Test.tsx which was setting a duplicate key prop.
