+++ apps/fluent-tester/src/RNTester/TestComponents/Button/ButtonFocusTest.tsx
@@ -4,9 +4,8 @@ import { Stack } from '@fluentui-react-native/stack';
 import * as React from 'react';
 import { findNodeHandle } from 'react-native';
 import { stackStyle } from '../Common/styles';
-import { Test, TestSection, PlatformStatus } from '../Test';
 
-const basicButton: React.FunctionComponent<{}> = () => {
+export const ButtonFocusTest: React.FunctionComponent<{}> = () => {
   const [state, setState] = React.useState({
     focused: false
   });
@@ -27,26 +26,5 @@ const basicButton: React.FunctionComponent<{}> = () => {
       <Button content={state.focused ? 'Focused' : 'Not Focused'} componentRef={buttonRef} accessibilityLabel="overridden button name" />
       <Button content="Click to focus" onClick={onFocus} tooltip="button tooltip" />
     </Stack>
-  );
-}
-
-const buttonSections: TestSection[] = [
-  {
-    name: 'Basic Button',
-    testID: BUTTON_TESTPAGE,
-    component: basicButton
-  }
-];
-
-export const ButtonFocusTest: React.FunctionComponent<{}> = () => {
-  const status: PlatformStatus = {
-    winStatus: 'beta',
-    iosStatus: 'experimental',
-    macosStatus: 'experimental',
-    androidStatus: 'experimental'
-  }
-
-  return (
-    <Test name="Button Test" description="No description." sections={buttonSections} status={status}></Test>
   );
 };
